change the file structure and the level of item_id
