+++ tabpfn_time_series/features/__init__.py
@@ -1,12 +1,13 @@
-from .feature_pipeline import (
+from .pipeline import (
     FeatureTransformer,
-    RunningIndexFeatureTransformer,
-    CalendarFeatureTransformer,
-    AutoSeasonalFeatureTransformer,
-    detrend,
+    ColumnConfig,
+    DefaultColumnConfig,
 )
+from .running_index import RunningIndexFeatureTransformer
+from .calendar_features import CalendarFeatureTransformer
+from .auto_seasonal import AutoSeasonalFeatureTransformer, detrend
 
-from .utils_pipeline import (
+from .utils import (
     train_test_split_time_series,
     from_autogluon_tsdf_to_df,
     from_df_to_autogluon_tsdf,
@@ -25,4 +26,6 @@ __all__ = [
     "from_df_to_autogluon_tsdf",
     "quick_mase_evaluation",
     "load_data",
+    "ColumnConfig",
+    "DefaultColumnConfig",
 ]
